requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
schedule==1.2.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
cryptography==41.0.7

